requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.10.0
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.2
openai==0.27.2
pandas==1.5.3
Pillow==9.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytrends==4.9.0
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
